% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{clr_norm}
\alias{clr_norm}
\title{Seurat CLR 算法在R中与在Rcpp的高效实现}
\usage{
clr_norm(counts, lang = c("R", "Rcpp"))
}
\arguments{
\item{mat}{稀疏矩阵}
}
\value{
返回标准化数据
}
\description{
Seurat CLR 算法在R中与在Rcpp的高效实现
}
