% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{regress_out_matrix}
\alias{regress_out_matrix}
\title{回归掉技术效应 和 细胞周期等函数
这里简单试下Seurat 的函数，用于学习 RegressOutMatrix
原来函数支持多个模型，这里只尝试 linear}
\usage{
regress_out_matrix(data, latent.data)
}
\arguments{
\item{data}{使用标准化矩阵}

\item{latent.data}{一个数据框，如 pbmc[\link{"percent.mt"}]}
}
\description{
回归掉技术效应 和 细胞周期等函数
这里简单试下Seurat 的函数，用于学习 RegressOutMatrix
原来函数支持多个模型，这里只尝试 linear
}
