% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{vst}
\alias{vst}
\title{Seurat 选择高可变基因算法
vst 使用的是原始数据，不是标准化的数据
本质上就是选取平均值高且}
\usage{
vst(counts, clip.max = "auto", loess.span = 0.3, nfeatures = 2000)
}
\description{
Seurat 选择高可变基因算法
vst 使用的是原始数据，不是标准化的数据
本质上就是选取平均值高且
}
